[REFACTOR/#26] 유틸 패키지 이동
